Update the exemple for good recon + take into account comments
