+++ pysap/numerics/utils.py
@@ -13,29 +13,6 @@ This module contains shortcuts to utility functions.
 
 # Package import
 from pysap.plugins.mri.reconstruct.utils import generate_operators
+from pysap.plugins.mri.reconstruct.utils import normalize_samples
 from pysap.plugins.mri.reconstruct.utils import convert_mask_to_locations
 from pysap.plugins.mri.reconstruct.utils import convert_locations_to_mask
-
-import numpy as np
-
-
-def normalize_samples(samples):
-    """
-    This function normalize the samples so it can be between [-0.5; 0.5[ for
-    the non-cartesian case
-    Parameters:
-    -----------
-        samples: np.ndarray
-            Unnormalized samples
-    Return:
-    -------
-        normalized_samples: np.ndarray
-            Same shape as the parameters but with values between [-0.5; 0.5[
-    """
-    samples_locations = np.copy(samples.astype('float'))
-    samples_locations[:, 0] /= 2 * np.abs(samples_locations[:, 0]).max()
-    samples_locations[:, 1] /= 2 * np.abs(samples_locations[:, 1]).max()
-    while samples_locations.max() == 0.5:
-        dim1, dim2 = np.where(samples_locations == 0.5)
-        samples_locations[dim1, dim2] = -0.5
-    return samples_locations
